fix: pr 리뷰 반영!
