+++ src/main/kotlin/rlc/webtoon/api/common/client/ApiResponse.kt
@@ -1,4 +1,4 @@
-package rlc.webtoon.api.common
+package rlc.webtoon.api.common.client
 
 data class ApiResponse<D>(
         val data: D
