+++ src/main/kotlin/rlc/webtoon/api/common/client/GlobalExceptionHandler.kt
@@ -1,4 +1,4 @@
-package rlc.webtoon.api.common
+package rlc.webtoon.api.common.client
 
 import io.jsonwebtoken.MalformedJwtException
 import org.springframework.dao.EmptyResultDataAccessException
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
-import java.lang.Exception
+import java.security.InvalidParameterException
 
 @ControllerAdvice
 class GlobalExceptionHandler {
@@ -15,6 +15,12 @@ class GlobalExceptionHandler {
     fun handle(e: ApiError) =
             run {
                 ResponseEntity(ErrorResponse(e.message), e.error.status)
+            }
+
+    @ExceptionHandler(value = [InvalidParameterException::class])
+    fun handle(e: InvalidParameterException) =
+            run {
+                ResponseEntity(ErrorResponse(e.message), HttpStatus.BAD_REQUEST)
             }
 
     @ExceptionHandler(value = [IllegalArgumentException::class])
@@ -29,6 +35,9 @@ class GlobalExceptionHandler {
                 ResponseEntity(ErrorResponse("존재하지 않는 데이터입니다."), HttpStatus.NOT_FOUND)
             }
 
+    /**
+     * jsonwebtoken 패키지안에 jwt 파싱할 경우 내부적으로 jwt 3 구성요소가 있는지 확인하고 없을 경우 밑에 예외가 발생된다.
+     */
     @ExceptionHandler(value = [MalformedJwtException::class])
     fun handle(e: MalformedJwtException) =
             run {
