+++ src/main/kotlin/rlc/webtoon/api/user/application/UserService.kt
@@ -3,13 +3,10 @@ package rlc.webtoon.api.user.application
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import rlc.webtoon.api.auth.application.AuthService
-import rlc.webtoon.api.common.ApiError
-import rlc.webtoon.api.common.Error
-import rlc.webtoon.api.common.PasswordEncoder
-import rlc.webtoon.api.user.domain.User
+import rlc.webtoon.api.common.client.ApiError
+import rlc.webtoon.api.common.client.Error
+import rlc.webtoon.api.common.util.PasswordEncoder
 import rlc.webtoon.api.user.infra.UserRepository
-import rlc.webtoon.api.user.presentation.dto.LoginRequest
-import rlc.webtoon.api.user.presentation.dto.LoginResponse
 import rlc.webtoon.api.user.presentation.dto.SignUpRequest
 
 
@@ -31,26 +28,6 @@ class UserService(
         userRepository.save(request.toUser(hashedPassword))
     }
 
-    @Transactional
-    fun login(request: LoginRequest): LoginResponse {
-        val user: User = userRepository.findByAccountId(request.accountId)
-
-        matchPassword(
-                rawPassword = request.password,
-                hashedPassword = user.password
-        )
-
-        val accessToken: String = authService.createAccessToken(user.accountId)
-        val refreshToken: String = authService.createRefreshToken(user.accountId)
-
-        user.addToken(refreshToken)
-
-        return LoginResponse(
-                accessToken = accessToken,
-                refreshToken = refreshToken,
-        )
-    }
-
     private fun checkDuplicateAccountId(accountId: String) {
 
         val isDuplicated: Boolean = userRepository.existsUserByAccountId(accountId)
@@ -61,12 +38,5 @@ class UserService(
 
     }
 
-    private fun matchPassword(rawPassword: String, hashedPassword: String) {
-
-        if (!passwordEncoder.match(rawPassword, hashedPassword)) {
-            throw ApiError(Error.INVALID_PASSWORD)
-        }
-
-    }
 
 }
