+++ src/test/kotlin/rlc/webtoon/api/common/PasswordEncoderTest.kt
@@ -1,14 +1,12 @@
 package rlc.webtoon.api.common
 
-import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.DisplayName
 
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ContextConfiguration
-import org.springframework.test.context.junit.jupiter.SpringJUnitConfig
+import rlc.webtoon.api.common.util.PasswordEncoder
 
 @SpringBootTest
 class PasswordEncoderTest @Autowired constructor(
