+++ src/test/kotlin/rlc/webtoon/api/user/application/UserServiceTest.kt
@@ -1,30 +1,22 @@
 package rlc.webtoon.api.user.application
 
-import org.apache.coyote.BadRequestException
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.*
 import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.MethodOrderer
-import org.junit.jupiter.api.Order
-
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestMethodOrder
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.dao.DuplicateKeyException
-import rlc.webtoon.api.common.PasswordEncoder
+import rlc.webtoon.api.auth.application.AuthService
 import rlc.webtoon.api.user.domain.User
 import rlc.webtoon.api.user.infra.UserRepository
 import rlc.webtoon.api.user.infra.UserTokenRepository
 import rlc.webtoon.api.user.presentation.dto.LoginRequest
 import rlc.webtoon.api.user.presentation.dto.LoginResponse
 import rlc.webtoon.api.user.presentation.dto.SignUpRequest
-import java.lang.IllegalArgumentException
 
 @SpringBootTest
 @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
 class UserServiceTest @Autowired constructor(
+        val authService: AuthService,
         val userService: UserService,
         val userRepository: UserRepository,
         val userTokenRepository: UserTokenRepository
@@ -36,7 +28,6 @@ class UserServiceTest @Autowired constructor(
         userRepository.deleteAllInBatch()
     }
 
-    @Order(1)
     @Test
     @DisplayName("사용자는 중복되지 않은 아이디와 비밀번호로 회원가입 한다.")
     fun signUp() {
@@ -55,7 +46,6 @@ class UserServiceTest @Autowired constructor(
         assertEquals(accountId, user.accountId)
     }
 
-    @Order(2)
     @Test
     @DisplayName("사용자가 중복된 아이디로 회원가입 할 경우 예외가 발생한다.")
     fun signUpWithDuplicatedAccountId() {
@@ -82,7 +72,6 @@ class UserServiceTest @Autowired constructor(
         assertEquals(exception.message, "이미 존재하는 사용자입니다.")
     }
 
-    @Order(3)
     @Test
     @DisplayName("사용자는 올바른 아이디와 비밀번호로 로그인 한다.")
     fun login() {
@@ -103,7 +92,7 @@ class UserServiceTest @Autowired constructor(
         userService.signUp(signUpRequest)
 
         // when
-        val response: LoginResponse = userService.login(loginRequest)
+        val response: LoginResponse = authService.login(loginRequest)
 
         // then
         assertNotNull(response.accessToken)
@@ -111,7 +100,6 @@ class UserServiceTest @Autowired constructor(
     }
 
 
-    @Order(4)
     @Test
     @DisplayName("사용자는 올바르지 않은 비밀번호로 로그인 할 경우 예외가 발생한다.")
     fun loginWithInvalidPassword() {
@@ -133,7 +121,7 @@ class UserServiceTest @Autowired constructor(
 
         // when then
         val exception = assertThrows(IllegalArgumentException::class.java) {
-            userService.login(loginRequest)
+            authService.login(loginRequest)
         }
 
         assertEquals(exception.message, "비밀번호가 일치하지 않습니다.")
